refactor(core): clean up circular dependencies
Moved code from `interfaces/i18n.ts` which was causing circular dependencies
