+++ goldens/circular-deps/packages.json
@@ -223,7 +223,6 @@
     "packages/core/src/render3/assert.ts",
     "packages/core/src/render3/interfaces/container.ts",
     "packages/core/src/render3/interfaces/node.ts",
-    "packages/core/src/render3/interfaces/i18n.ts",
     "packages/core/src/render3/interfaces/view.ts",
     "packages/core/src/di/injector.ts",
     "packages/core/src/di/r3_injector.ts",
@@ -240,7 +239,6 @@
     "packages/core/src/render3/assert.ts",
     "packages/core/src/render3/interfaces/container.ts",
     "packages/core/src/render3/interfaces/node.ts",
-    "packages/core/src/render3/interfaces/i18n.ts",
     "packages/core/src/render3/interfaces/view.ts",
     "packages/core/src/metadata.ts",
     "packages/core/src/di.ts",
@@ -264,7 +262,6 @@
     "packages/core/src/render3/assert.ts",
     "packages/core/src/render3/interfaces/container.ts",
     "packages/core/src/render3/interfaces/node.ts",
-    "packages/core/src/render3/interfaces/i18n.ts",
     "packages/core/src/render3/interfaces/view.ts",
     "packages/core/src/render3/interfaces/definition.ts",
     "packages/core/src/core.ts",
@@ -971,13 +968,11 @@
   [
     "packages/core/src/render3/interfaces/container.ts",
     "packages/core/src/render3/interfaces/node.ts",
-    "packages/core/src/render3/interfaces/i18n.ts",
     "packages/core/src/render3/interfaces/view.ts"
   ],
   [
     "packages/core/src/render3/interfaces/definition.ts",
     "packages/core/src/render3/interfaces/node.ts",
-    "packages/core/src/render3/interfaces/i18n.ts",
     "packages/core/src/render3/interfaces/view.ts"
   ],
   [
@@ -985,23 +980,13 @@
     "packages/core/src/render3/interfaces/view.ts"
   ],
   [
-    "packages/core/src/render3/interfaces/i18n.ts",
-    "packages/core/src/render3/interfaces/node.ts"
-  ],
-  [
-    "packages/core/src/render3/interfaces/i18n.ts",
+    "packages/core/src/render3/interfaces/node.ts",
     "packages/core/src/render3/interfaces/view.ts"
   ],
   [
-    "packages/core/src/render3/interfaces/i18n.ts",
+    "packages/core/src/render3/interfaces/node.ts",
     "packages/core/src/render3/interfaces/view.ts",
-    "packages/core/src/render3/interfaces/node.ts"
-  ],
-  [
-    "packages/core/src/render3/interfaces/i18n.ts",
-    "packages/core/src/render3/interfaces/view.ts",
-    "packages/core/src/render3/interfaces/query.ts",
-    "packages/core/src/render3/interfaces/node.ts"
+    "packages/core/src/render3/interfaces/query.ts"
   ],
   [
     "packages/core/src/render3/interfaces/query.ts",
